gdb-common: Add --with-liblzma-prefix in xz PACKAGECONFIG.
This fixes an arch linux gdb configure error with liblzma installed.

Without the --with-liblzma-prefix option, when configuring the target
gdb, gdb configure may find the native liblzma (on a non-multiarch
system) and try to use it, which gives a configure error.  We already
use --with-libexpat-prefix for expat, we just need to do the same for
liblzma.

This was reported here:
https://github.com/sifive/meta-sifive/issues/34
https://github.com/sifive/freedom-u-sdk/issues/148

Signed-off-by: Jim Wilson <jim.wilson.gcc@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>