unzip/zip: mark all patches as non-upstreamable
While upstream is not technically dead as of today [1], it has not made
a release in over a decade, and there is no source code repo to
rebase the patches on. If/when a new release ever comes, or someone
takes over the maintenance, we can reassess the situation with these patches.

[1] Discussion in https://sourceforge.net/p/infozip/bugs/53/

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>